requirements.txt
arrow==1.3.0
astroid==3.1.0
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
dill==0.3.8
fonttools==4.51.0
idna==3.7
isort==5.13.2
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
Pygments==2.17.2
pylint==3.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
text-unidecode==1.3
threadpoolctl==3.4.0
tomli==2.0.1
tomlkit==0.12.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
